+++ rust/operator-binary/src/controller/mod.rs
@@ -643,7 +643,7 @@ mod tests {
               uid: e6ac237d-a6d4-43a1-8135-f36506110912
             spec:
               image:
-                productVersion: 3.2.2
+                productVersion: 3.3.1
               clusterConfig:
                 credentialsSecretName: airflow-admin-credentials
                 metadataDatabase: