+++ rust/operator-binary/src/controller/validate.rs
@@ -454,9 +454,9 @@ mod tests {
     fn test_resolved_product_image()
     -> stackable_operator::commons::product_image_selection::ResolvedProductImage {
         stackable_operator::commons::product_image_selection::ResolvedProductImage {
-            product_version: "3.0.6".to_string(),
-            app_version_label_value: "3.0.6".parse().expect("valid label value"),
-            image: "oci.example.org/sdp/airflow:3.0.6-stackable0.0.0-dev".to_string(),
+            product_version: "3.3.1".to_string(),
+            app_version_label_value: "3.3.1".parse().expect("valid label value"),
+            image: "oci.example.org/sdp/airflow:3.3.1-stackable0.0.0-dev".to_string(),
             image_pull_policy: "IfNotPresent".to_string(),
             pull_secrets: None,
         }
@@ -470,7 +470,7 @@ mod tests {
           name: airflow
         spec:
           image:
-            productVersion: 3.1.6
+            productVersion: 3.3.1
           clusterConfig:
             loadExamples: false
             exposeConfig: false
@@ -570,7 +570,7 @@ mod tests {
           name: airflow
         spec:
           image:
-            productVersion: 3.1.6
+            productVersion: 3.3.1
           clusterConfig:
             loadExamples: false
             exposeConfig: false
@@ -651,7 +651,7 @@ mod tests {
           name: airflow
         spec:
           image:
-            productVersion: 3.1.6
+            productVersion: 3.3.1
           clusterConfig:
             loadExamples: false
             exposeConfig: false